% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_rqda.R
\name{rql_import_rqda}
\alias{rql_import_rqda}
\title{Import data from RQDA to Requal}
\usage{
rql_import_rqda(rqda_file, requal_file)
}
\arguments{
\item{rqda_file}{Path to RQDA database}

\item{requal_file}{Path to requal database to be created by import, it must not exist}
}
\description{
Import data from RQDA database to requal. It uses tables
RQDA tables "project", "source" (documents), "freecode" (codes) 
and "coding" (coded segments), "cases", "caselinkage", "codecat" (categories).
Memos from "source", "freecode", "cases" and "codecat" are used as descriptions. 
Memos from "coding" are stored as memos.
}
