% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_rqda.R
\name{import_rqda}
\alias{import_rqda}
\title{Import data from RQDA to Requal}
\usage{
import_rqda(rqda_file, requal_connection)
}
\arguments{
\item{rqda_file}{Path to RQDA database}

\item{requal_connection}{Connection to REQUAL database}
}
\description{
Import data from RQDA database to requal. Only import from
RQDA tables "project", "source" (documents), "freecode" (codes) 
and "coding" (coded segments) is implemented.
Memos from "source" and "freecode" are used as descriptions. 
Memos from "coding" are stored as memos.
}
