% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(
  mode = "local",
  dbhost = "localhost",
  dbport = 5432,
  dbname = NULL,
  dbusername = NULL,
  dbpassword = NULL,
  credentials_path = NULL,
  credentials_pass = NULL,
  uiPattern = "/",
  ...
)
}
\arguments{
\item{mode}{Mode of application (local or server)}

\item{dbhost}{Host of DB (for server use only)}

\item{dbport}{Port of DB (for server use only)}

\item{dbname}{Name of DB (for server use only)}

\item{dbusername}{Username (for server use only)}

\item{dbpassword}{Password to DB (for server use only)}

\item{credentials_path}{Path to credentials DB (for server use only)}

\item{credentials_pass}{Password to credentials DB (for server use only)}

\item{uiPattern}{A regular expression that will be applied to each \code{GET}
request to determine whether the \code{ui} should be used to handle the
request. Note that the entire request path must match the regular
expression in order for the match to be considered successful.}

\item{...}{arguments to pass to golem_opts.
See `?golem::get_golem_options` for more details.}
}
\description{
Run the Shiny Application
}
